ENH: Add to_latex() method to Series (#16180) (#16465)
* ENH: Add to_latex() method to Series (#16180)

This changeset adds _repr_latex_ to the Series class and moves
the to_latex() method from the DataFrame class to the NDFrame
class.

* Add Series to_latex test

* Move _repr_latex_ to NDFrame

Streamline things a bit by moving _repr_latex_ methods out of
the Series and DataFrame classes

* DOC: Added versionchanged

(cherry picked from commit 6a6227d19b6aa522d6ae5362da0589a45913d7bd)
